fixed the bugs in the test and js file
1- I followed the pattern from one of the test files
that used '$(document).ready(() => {' inside the tests,
turns out those tests aren't running. So the problem
with my tests was they weren't running.

2- On why the js stopped working: there is a regex in line
104 that was testing only for digits in the pathname, but
this is the pathname we get:
'/casa_cases/CINA-07-1003/emancipation'

because of this we returned and the code didn't run,
that's why this also fails in main.